requirements.txt
llama-cpp-python
langchain
sentence-transformers
faiss-cpu
pypdf
python-docx
streamlit
fastapi
uvicorn
